requirements.txt
asgiref==3.5.2
boto3==1.17.58
botocore==1.20.58
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.1.0
cloudinary==1.29.0
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.9
django-admin-interface==0.19.1
django-admin-thumbnails==0.2.6
django-autoslug==1.9.8
django-autoslug-field==0.2.3
django-ckeditor==5.9.0
django-cloudinary-storage==0.3.0
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-countries==7.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-embed-video==1.4.4
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-import-export==2.5.0
django-js-asset==1.2.2
django-recaptcha==3.0.0
django-storages==1.11.1
django-tawkto==0.3
django-tinymce==3.4.0
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.7.1
gunicorn==20.1.0
html5lib==1.1
idna==2.10
jmespath==0.10.0
MarkupPy==1.14
mysql-client==0.0.1
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
Pillow==9.2.0
pipenv==2022.7.4
platformdirs==2.5.2
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
Pyphen==0.10.0
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2020.1
PyYAML==5.4.1
requests==2.28.1
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.2
tablib==3.0.0
text-unidecode==1.3
tinycss2==1.1.0
twilio==6.32.0
tzdata==2022.1
urllib3==1.26.10
virtualenv==20.15.1
virtualenv-clone==0.5.7
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
